fix judge of fail job (#767)

+++ scripts/save.py
@@ -477,7 +477,7 @@ def parse_logs(args):
             # check_results and send alarm email
             if job_info["index"] == 1:  # speed
                 print_machine_type = machine_type_to_print(run_machine_type)
-                if int(result) == 0 or os.getenv('job_fail_flag') == 1:
+                if float(result) == 0 or os.getenv('job_fail_flag') == 1:
                     print('job_fail_flag:{}'.format(os.getenv('job_fail_flag')))
                     FAIL_LIST.append([job_info["model_name"], print_machine_type])
                     outlier = 1
